Switch to a more standard way of loading pseudo-steps
This is how Declarative works - it avoids needing to do any
whitelisting magic around calling other steps, which seems to have
been the cause for the worst of the whitelistings. More refinement is
definitely still needed, and honestly, I don't know if this works,
because I don't have a k8s env locally. =)