Move GH functions to JXDSLUtils, clean up code some
This all *should* end up being changed to use the github-api library,
ideally with the backoff-retry logic for API rate limits that's used
in github-branch-source as well, but there's a core problem with that
currently: the token the plugin's expecting to use is a GitHub Apps
token, but github-api only supports OAuth tokens or username +
personal access tokens. So either we need to add GitHub Apps token
support to github-api (definitely doable), or we need to change the
expected authentication mechanism here.