Initial validation support.
Note that the example we actually test for is built-in from
Declarative's own validation of `Describable`s, but I've also added
hooks for more detailed validation being added to the various *DSL
classes.

Also added an easy way to get the appropriate *Arguments descriptor
for a given *DSL class (currently unused but will be useful for
further validation) and got rid of a bunch of redundant whitelist
extensions. More tests and probably some more detailed validation to
come soon.